requirements.txt
APScheduler==3.11.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy==1.13.0
mypy-extensions==1.0.0
pyloggerhelper==0.0.4
pyproxypattern==0.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
schema_entry==0.2.1
structlog==24.4.0
tomli==2.2.1
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
